Most Functionality + Messing Around
Got most of the functionality down. Have a slash command that takes user input and will create/post a message in a thread with the input from the user.
Also have attempts at making a context menu and modal and other stuff that was abandoned.
Main important changes are in spoiler.js