New Book Button, New Thread Validation, Author's Username
Pulled out the New Book option from the list and made it a button.
Added validation to ensure that if you select New Button and type in an existing thread, it just posts that spoiler in the existing thread
Changed the name associated with the spoiler post to be the author of the message as opposed to the individual who triggered the command. And changed it from the guild name to the username because some people like to change their server username all the time